extract authConfig to a new file
